# Conflicts:
#	src/helpers/error-messages.js
#	src/services/connector-service.js

+++ src/services/catalog-service.js
@@ -44,8 +44,13 @@ const updateCatalogItem = async function (id, data, user, isCLI, transaction) {
   await Validator.validate(data, Validator.schemas.catalogItemUpdate);
 
   const where = isCLI
-    ? {id: id}
-    : {id: id, userId: user.id};
+    ? {
+         id: id
+      }
+    : {
+         id: id,
+         userId: user.id
+      };
 
   data.id = id;
   await _updateCatalogItem(data, where, transaction);
@@ -55,7 +60,7 @@ const updateCatalogItem = async function (id, data, user, isCLI, transaction) {
 
 const listCatalogItems = async function (user, isCLI, transaction) {
   const where = isCLI
-    ? {[Op.or]: [{category: {[Op.ne]: 'SYSTEM'}}, {category: null}]}
+    ? {}
     : {
       [Op.or]: [{userId: user.id}, {userId: null}],
       [Op.or]: [{category: {[Op.ne]: 'SYSTEM'}}, {category: null}]
@@ -73,7 +78,7 @@ const listCatalogItems = async function (user, isCLI, transaction) {
 
 const getCatalogItem = async function (id, user, isCLI, transaction) {
   const where = isCLI
-    ? {[Op.or]: [{category: {[Op.ne]: 'SYSTEM'}}, {category: null}]}
+    ? {}
     : {
       [Op.or]: [{userId: user.id}, {userId: null}],
       [Op.or]: [{category: {[Op.ne]: 'SYSTEM'}}, {category: null}]
@@ -91,9 +96,22 @@ const getCatalogItem = async function (id, user, isCLI, transaction) {
 };
 
 const deleteCatalogItem = async function (id, user, isCLI, transaction) {
+
   const where = isCLI
-    ? {id: id}
-    : {userId: user.id, id: id};
+    ? {
+         id: id
+      }
+    : {
+        userId: user.id,
+        id: id
+      };
+
+  const item = await _checkIfItemExists(where, transaction);
+
+  if (item.category == "SYSTEM"){
+    throw new Errors.ValidationError(AppHelper.formatMessage(ErrorMessages.SYSTEM_CATALOG_ITEM_DELETE, id));
+  }
+
   const affectedRows = await CatalogItemManager.delete(where, transaction);
   if (affectedRows === 0) {
     throw new Errors.NotFoundError(AppHelper.formatMessage(ErrorMessages.INVALID_CATALOG_ITEM_ID, id));
@@ -176,6 +194,13 @@ const _createCatalogItem = async function (data, user, transaction) {
   };
 
   catalogItem = AppHelper.deleteUndefinedFields(catalogItem);
+
+  if (catalogItem.registryId) {
+      const registry = await RegistryManager.findOne({id: catalogItem.registryId}, transaction);
+      if (!registry) {
+          throw new Errors.NotFoundError(AppHelper.formatMessage(ErrorMessages.INVALID_REGISTRY_ID, data.registryId));
+      }
+  }
 
   return await CatalogItemManager.create(catalogItem, transaction);
 };
@@ -264,6 +289,11 @@ const _updateCatalogItem = async function (data, where, transaction) {
   }
 
   const item = await _checkIfItemExists(where, transaction);
+
+  if (item.category == "SYSTEM"){
+    throw new Errors.ValidationError(AppHelper.formatMessage(ErrorMessages.SYSTEM_CATALOG_ITEM_UPDATE, data.id));
+  }
+
   await _checkForDuplicateName(data.name, item, transaction);
   await CatalogItemManager.update(where, catalogItem, transaction);
 };
